requirements.txt
fastapi
uvicorn
python-dotenv
openai
sentence-transformers
faiss-cpu
pydantic
pdfplumber